renamed the package passed test
