+++ src/main/java/com/vitruv/methodologist/MethodologistApplication.java
@@ -1,0 +1,15 @@
+package com.vitruv.methodologist;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.web.bind.annotation.RestController;
+
+@SpringBootApplication
+@RestController
+@EnableScheduling
+public class MethodologistApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(MethodologistApplication.class, args);
+    }
+}
