+++ src/main/java/com/vitruv/methodologist/ResponseTemplateDto.java
@@ -1,0 +1,15 @@
+package com.vitruv.methodologist;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+public class ResponseTemplateDto<T> {
+    private T data;
+    private String message;
+}
