+++ src/main/java/com/vitruv/methodologist/general/controller/responsedto/LatestVersionResponse.java
@@ -1,0 +1,13 @@
+package com.vitruv.methodologist.general.controller.responsedto;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class LatestVersionResponse {
+    private Boolean forceUpdate;
+    private String version;
+}
