+++ src/main/java/com/vitruv/methodologist/general/model/repository/VersioningRepository.java
@@ -1,0 +1,11 @@
+package com.vitruv.methodologist.general.model.repository;
+
+
+import com.vitruv.methodologist.general.model.Versioning;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface VersioningRepository extends CrudRepository<Versioning, Long> {
+    Optional<Versioning> findTopByAppNameOrderByIdDesc(String name);
+}
