+++ src/main/java/com/vitruv/methodologist/messages/Error.java
@@ -1,0 +1,5 @@
+package com.vitruv.methodologist.messages;
+
+public class Error {
+    public static final String CLIENT_NOT_FOUND_ERROR =  "Client not found";
+}
