+++ src/main/java/com/vitruv/methodologist/messages/Message.java
@@ -1,0 +1,6 @@
+package com.vitruv.methodologist.messages;
+
+public class Message {
+    public static final String LOGIN_USER_SUCCESSFULLY = "User successfully logged in";
+    public static final String SIGNUP_USER_SUCCESSFULLY = "User successfully signed up";
+}
