+++ src/main/java/com/vitruv/methodologist/user/controller/dto/response/UserResponse.java
@@ -1,0 +1,14 @@
+package com.vitruv.methodologist.user.controller.dto.response;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserResponse {
+    private Long id;
+    private String firstName;
+    private String lastName;
+}
