+++ src/main/java/com/vitruv/methodologist/user/mapper/UserMapper.java
@@ -1,0 +1,8 @@
+package com.vitruv.methodologist.user.mapper;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring")
+public interface UserMapper {
+}
