+++ src/main/java/com/vitruv/methodologist/user/model/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.vitruv.methodologist.user.model.repository;
+
+import com.vitruv.methodologist.user.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+}
