+++ src/main/java/com/vitruv/methodologist/user/service/UserService.java
@@ -1,0 +1,18 @@
+package com.vitruv.methodologist.user.service;
+
+import com.vitruv.methodologist.user.mapper.UserMapper;
+import com.vitruv.methodologist.user.model.repository.UserRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+public class UserService {
+    private final UserMapper userMapper;
+    private final UserRepository userRepository;
+
+    public UserService(UserMapper userMapper, UserRepository userRepository) {
+        this.userMapper = userMapper;
+        this.userRepository = userRepository;
+    }
+}
