--- src/main/java/com/vituv/methodologist/MethodologistApplication.java
@@ -1,15 +1,0 @@
-package com.vituv.methodologist;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-@RestController
-@EnableScheduling
-public class MethodologistApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(MethodologistApplication.class, args);
-    }
-}
