--- src/main/java/com/vituv/methodologist/ResponseTemplateDto.java
@@ -1,15 +1,0 @@
-package com.vituv.methodologist;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@Builder
-@AllArgsConstructor
-public class ResponseTemplateDto<T> {
-    private T data;
-    private String message;
-}
