--- src/main/java/com/vituv/methodologist/exception/NotFoundException.java
@@ -1,9 +1,0 @@
-package com.vituv.methodologist.exception;
-
-public class NotFoundException extends RuntimeException {
-    public static final String messageTemplate = "%s not found!";
-
-    public NotFoundException(String objectName) {
-        super(String.format(messageTemplate, objectName));
-    }
-}
