--- src/main/java/com/vituv/methodologist/general/controller/responsedto/LatestVersionResponse.java
@@ -1,13 +1,0 @@
-package com.vituv.methodologist.general.controller.responsedto;
-
-import lombok.*;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class LatestVersionResponse {
-    private Boolean forceUpdate;
-    private String version;
-}
