--- src/main/java/com/vituv/methodologist/general/mapper/VersioningMapper.java
@@ -1,11 +1,0 @@
-package com.vituv.methodologist.general.mapper;
-
-import com.vituv.methodologist.general.controller.responsedto.LatestVersionResponse;
-import com.vituv.methodologist.general.model.Versioning;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring")
-public interface VersioningMapper {
-    LatestVersionResponse toLatestVersionResponse(Versioning versioning);
-}
