--- src/main/java/com/vituv/methodologist/general/model/Versioning.java
@@ -1,34 +1,0 @@
-package com.vituv.methodologist.general.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class Versioning {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotNull
-    @NotBlank
-    private String appName;
-
-    @NotNull
-    @NotBlank
-    private String version;
-
-    @NotNull
-    private Boolean forceUpdate;
-}
