--- src/main/java/com/vituv/methodologist/messages/Error.java
@@ -1,5 +1,0 @@
-package com.vituv.methodologist.messages;
-
-public class Error {
-    public static final String CLIENT_NOT_FOUND_ERROR =  "Client not found";
-}
