--- src/main/java/com/vituv/methodologist/messages/Message.java
@@ -1,6 +1,0 @@
-package com.vituv.methodologist.messages;
-
-public class Message {
-    public static final String LOGIN_USER_SUCCESSFULLY = "User successfully logged in";
-    public static final String SIGNUP_USER_SUCCESSFULLY = "User successfully signed up";
-}
