--- src/main/java/com/vituv/methodologist/user/controller/dto/response/UserResponse.java
@@ -1,14 +1,0 @@
-package com.vituv.methodologist.user.controller.dto.response;
-
-import lombok.*;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserResponse {
-    private Long id;
-    private String firstName;
-    private String lastName;
-}
