--- src/main/java/com/vituv/methodologist/user/mapper/UserMapper.java
@@ -1,8 +1,0 @@
-package com.vituv.methodologist.user.mapper;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring")
-public interface UserMapper {
-}
