--- src/main/java/com/vituv/methodologist/user/model/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.vituv.methodologist.user.model.repository;
-
-import com.vituv.methodologist.user.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
