--- src/main/java/com/vituv/methodologist/user/service/UserService.java
@@ -1,18 +1,0 @@
-package com.vituv.methodologist.user.service;
-
-import com.vituv.methodologist.user.mapper.UserMapper;
-import com.vituv.methodologist.user.model.repository.UserRepository;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Service
-@Slf4j
-public class UserService {
-    private final UserMapper userMapper;
-    private final UserRepository userRepository;
-
-    public UserService(UserMapper userMapper, UserRepository userRepository) {
-        this.userMapper = userMapper;
-        this.userRepository = userRepository;
-    }
-}
