+++ src/main/resources/db/migration/V1__START.sql
@@ -1,124 +1,13 @@
-CREATE TABLE asset
-(
-    id                 BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    name               VARCHAR(255)                            NOT NULL,
-    symbol             VARCHAR(255)                            NOT NULL,
-    description        VARCHAR(4000)                           NOT NULL,
-    risk_type          VARCHAR(255)                            NOT NULL,
-    duration_type      VARCHAR(255)                            NOT NULL,
-    asset_category_id  BIGINT                                  NOT NULL,
-    is_active          BOOLEAN                                 NOT NULL,
-    created_at         TIMESTAMP WITHOUT TIME ZONE,
-    updated_at         TIMESTAMP WITHOUT TIME ZONE,
-    icon_address       VARCHAR(255)                            NOT NULL,
-    provider           VARCHAR(255)                            NOT NULL,
-    price_provider     VARCHAR(255)                            NOT NULL,
-    external_price_id  VARCHAR(255),
-    price              DOUBLE PRECISION                        NOT NULL,
-    last_price_updated TIMESTAMP WITHOUT TIME ZONE             NOT NULL,
-    priority           INTEGER                                 NOT NULL,
-    CONSTRAINT pk_asset PRIMARY KEY (id)
-);
-
-CREATE TABLE asset_category
-(
-    id           BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    fa_name      VARCHAR(255)                            NOT NULL,
-    description  VARCHAR(255)                            NOT NULL,
-    icon_address VARCHAR(255)                            NOT NULL,
-    CONSTRAINT pk_assetcategory PRIMARY KEY (id)
-);
-
-CREATE TABLE asset_suggestion
-(
-    id                 BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    percentage         DOUBLE PRECISION                        NOT NULL,
-    asset_id           BIGINT                                  NOT NULL,
-    person_category_id BIGINT                                  NOT NULL,
-    CONSTRAINT pk_assetsuggestion PRIMARY KEY (id)
-);
-
-CREATE TABLE card
-(
-    id           BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    user_id      BIGINT                                  NOT NULL,
-    name         VARCHAR(255),
-    card_no      BIGINT                                  NOT NULL,
-    bank_name    VARCHAR(255),
-    default_card BOOLEAN                                 NOT NULL,
-    created_at   TIMESTAMP WITHOUT TIME ZONE,
-    removed_at   TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_card PRIMARY KEY (id)
-);
-
-CREATE TABLE ord
-(
-    id             BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    user_id        BIGINT                                  NOT NULL,
-    asset_id       BIGINT                                  NOT NULL,
-    payment_id     BIGINT                                  NOT NULL,
-    status         VARCHAR(255)                            NOT NULL,
-    amount         BIGINT                                  NOT NULL,
-    unit_price_irt BIGINT,
-    unit_price_usd DOUBLE PRECISION,
-    quantity       DOUBLE PRECISION,
-    margin_irt     BIGINT,
-    margin_usd     DOUBLE PRECISION,
-    created_at     TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_ord PRIMARY KEY (id)
-);
-
-CREATE TABLE payment
-(
-    id           BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    receipt_id   VARCHAR(255),
-    tracing_code BIGINT                                  NOT NULL,
-    status       VARCHAR(255)                            NOT NULL,
-    type         VARCHAR(255)                            NOT NULL,
-    created_at   TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_payment PRIMARY KEY (id)
-);
-
-CREATE TABLE person_category
-(
-    id           BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    name         VARCHAR(255)                            NOT NULL,
-    rank         BIGINT                                  NOT NULL,
-    description  VARCHAR(4000)                           NOT NULL,
-    icon_address VARCHAR(255)                            NOT NULL,
-    CONSTRAINT pk_personcategory PRIMARY KEY (id)
-);
-
-CREATE TABLE user_asset
-(
-    id                BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    user_id           BIGINT                                  NOT NULL,
-    asset_id          BIGINT                                  NOT NULL,
-    quantity          DOUBLE PRECISION                        NOT NULL,
-    total_buy_amount  BIGINT                                  NOT NULL,
-    total_sell_amount BIGINT                                  NOT NULL,
-    created_at        TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_userasset PRIMARY KEY (id)
-);
-
 CREATE TABLE usr
 (
-    id                          BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    email                       VARCHAR(255),
-    person_category_id          BIGINT,
-    wallet_balance              DOUBLE PRECISION                        NOT NULL,
-    username                    VARCHAR(255),
-    first_name                  VARCHAR(255),
-    last_name                   VARCHAR(255),
-    referral_code               VARCHAR(255),
-    registered_referral_code    VARCHAR(255),
-    kyc                         BOOLEAN                                 NOT NULL,
-    kyc_session_id              VARCHAR(255),
-    phone_number                VARCHAR(255)                            NOT NULL,
-    created_at                  TIMESTAMP WITHOUT TIME ZONE,
-    removed_at                  TIMESTAMP WITHOUT TIME ZONE,
-    verification_code           VARCHAR(255),
-    verification_code_sent_time TIMESTAMP WITHOUT TIME ZONE,
+    id         BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    email      VARCHAR(255),
+    role_type  VARCHAR(255)                            NOT NULL,
+    username   VARCHAR(255),
+    first_name VARCHAR(255),
+    last_name  VARCHAR(255),
+    created_at TIMESTAMP WITHOUT TIME ZONE,
+    removed_at TIMESTAMP WITHOUT TIME ZONE,
     CONSTRAINT pk_usr PRIMARY KEY (id)
 );
 
@@ -130,39 +19,3 @@ CREATE TABLE versioning
     force_update BOOLEAN                                 NOT NULL,
     CONSTRAINT pk_versioning PRIMARY KEY (id)
 );
-
-ALTER TABLE payment
-    ADD CONSTRAINT uc_payment_tracingcode UNIQUE (tracing_code);
-
-ALTER TABLE usr
-    ADD CONSTRAINT uc_usr_phonenumber UNIQUE (phone_number);
-
-ALTER TABLE asset_suggestion
-    ADD CONSTRAINT FK_ASSETSUGGESTION_ON_ASSET FOREIGN KEY (asset_id) REFERENCES asset (id);
-
-ALTER TABLE asset_suggestion
-    ADD CONSTRAINT FK_ASSETSUGGESTION_ON_PERSON_CATEGORY FOREIGN KEY (person_category_id) REFERENCES person_category (id);
-
-ALTER TABLE asset
-    ADD CONSTRAINT FK_ASSET_ON_ASSET_CATEGORY FOREIGN KEY (asset_category_id) REFERENCES asset_category (id);
-
-ALTER TABLE card
-    ADD CONSTRAINT FK_CARD_ON_USER FOREIGN KEY (user_id) REFERENCES usr (id);
-
-ALTER TABLE ord
-    ADD CONSTRAINT FK_ORD_ON_ASSET FOREIGN KEY (asset_id) REFERENCES asset (id);
-
-ALTER TABLE ord
-    ADD CONSTRAINT FK_ORD_ON_PAYMENT FOREIGN KEY (payment_id) REFERENCES payment (id);
-
-ALTER TABLE ord
-    ADD CONSTRAINT FK_ORD_ON_USER FOREIGN KEY (user_id) REFERENCES usr (id);
-
-ALTER TABLE user_asset
-    ADD CONSTRAINT FK_USERASSET_ON_ASSET FOREIGN KEY (asset_id) REFERENCES asset (id);
-
-ALTER TABLE user_asset
-    ADD CONSTRAINT FK_USERASSET_ON_USER FOREIGN KEY (user_id) REFERENCES usr (id);
-
-ALTER TABLE usr
-    ADD CONSTRAINT FK_USR_ON_PERSON_CATEGORY FOREIGN KEY (person_category_id) REFERENCES person_category (id);
