+++ src/test/java/com/vitruv/methodologist/MethodologistApplicationTests.java
@@ -5,9 +5,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class MethodologistApplicationTests {
-
 	@Test
-	void contextLoads() {
-	}
-
+	void contextLoads() {}
 }
